moved functions to separate files and imported; [TODO]: pls test kay wala diri ang db sakoa
